Started cleaning up and fixing bugs.
