+++ src/parser/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_subdirectory(sparqlParser)
 
 add_library(rdfEscaping RdfEscaping.h RdfEscaping.cpp)
-target_link_libraries(rdfEscaping ${ICU_LIBRARIES})
+target_link_libraries(rdfEscaping absl::flat_hash_map ${ICU_LIBRARIES})
 
 add_library(parser
         SparqlParser.h SparqlParser.cpp
