+++ src/util/Random.h
@@ -30,8 +30,8 @@ class FastRandomIntGenerator {
   FastRandomIntGenerator() {
     // Randomly initialize the shuffleTable
     std::random_device seeder{};
-    for (size_t i = 0; i <= _shuffleTable.size(); ++i) {
-      _shuffleTable[i] = seeder();
+    for (auto& el : _shuffleTable) {
+      el = seeder();
     }
   }
